refactor(ui): restructure run info tests as run-page.test.tsx
Match the deployment-page.test.tsx structure: entity-level describe with
per-tab nested describes and per-test inline wrapper composition.
